filtering: the clamp under the name it now has
Merging main brings the exported ClampResponseSize and the typed
MaxQueryFilterLimit, and neither arrives as a conflict — the rename is in a file
this branch does not touch, and the const-to-var change is a type error rather
than a textual one. Both surface only at build time.

- sql.go binds through ClampResponseSize, which is the same function under the
  name it now has. Nothing about the ordering changes: the clamp still lands
  before the narrowing to int32, and an explicit zero is still left alone.
- sql_test.go spells int32(MaxQueryFilterLimit) where it was relying on the
  untyped constant to take the field's type, and drops the uint16 conversion
  that a typed ceiling makes redundant.

A page size at the ceiling is now a page size at whatever ceiling the consumer
set, since Bind reads the var rather than a value fixed when the package
compiled. That is the point of making it settable, and Bind needed no change to
follow it.

+++ filtering/sql_test.go
@@ -76,15 +76,15 @@ func TestBind(T *testing.T) {
 		// answer.
 		args := Bind(&QueryFilter{MaxResponseSize: new(uint16(60_000))})
 
-		test.EqOp(t, sql.NullInt32{Int32: MaxQueryFilterLimit, Valid: true}, args.ResultLimit)
+		test.EqOp(t, sql.NullInt32{Int32: int32(MaxQueryFilterLimit), Valid: true}, args.ResultLimit)
 	})
 
 	T.Run("with a page size at the ceiling", func(t *testing.T) {
 		t.Parallel()
 
-		args := Bind(&QueryFilter{MaxResponseSize: new(uint16(MaxQueryFilterLimit))})
+		args := Bind(&QueryFilter{MaxResponseSize: new(MaxQueryFilterLimit)})
 
-		test.EqOp(t, sql.NullInt32{Int32: MaxQueryFilterLimit, Valid: true}, args.ResultLimit)
+		test.EqOp(t, sql.NullInt32{Int32: int32(MaxQueryFilterLimit), Valid: true}, args.ResultLimit)
 	})
 
 	T.Run("preserving an explicit zero page size", func(t *testing.T) {
